PSET7/numb3rs/test_numb3rs.py
import numb3rs  # Import the numb3rs module, which contains the validate function

# Test cases for valid IPv4 addresses
def test_valid_ipv4_addresses():
    # Test a valid IPv4 address
    assert numb3rs.validate("127.0.0.1") == True  # Loopback address, valid
    assert numb3rs.validate("255.255.255.255") == True  # Broadcast address, valid
    assert numb3rs.validate("140.247.235.144") == True  # Another valid IPv4 address

# Test cases for invalid IPv4 addresses
def test_invalid_ipv4_addresses():
    # Test an address with an out-of-range number
    assert numb3rs.validate("256.0.0.1") == False  # Invalid because '256' is outside the valid range (0-255)
    
    # Test an address with multiple numbers out of range
    assert numb3rs.validate("64.128.256.512") == False  # Invalid because '256' and '512' are out of range
    
    # Test an address with fewer than 4 components (missing parts)
    assert numb3rs.validate("8.8.8") == False  # Invalid, not enough components
    
    # Test an address with too many components (more than 4)
    assert numb3rs.validate("10.10.10.10.10") == False  # Invalid, more than 4 components
    
    # Test an address that is an IPv6 address (not supported by the validation logic)
    assert numb3rs.validate("2001:0db8:85a3:0000:0000:8a2e:0370:7334") == False  # Invalid IPv6 address
    
    # Test an address that is not an IP address at all
    assert numb3rs.validate("cat") == False  # Invalid, not a valid IP address

# Run the test cases when the script is executed
if __name__ == '__main__':
    test_valid_ipv4_addresses()  # Run the tests for valid IPv4 addresses
    test_invalid_ipv4_addresses()  # Run the tests for invalid IPv4 addresses
